# CLD_bound: speed up by precomputing double polynomials