This is a cheap 10% speedup for the fmpz_poly_factor benchmark.

Before:

```
P1 has 36 factors: 15 ms
P2 has 12 factors: 33 ms
P3 has 16 factors: 22 ms
P4 has 2 factors: 259 ms
P5 has 1 factors: 8 ms
P6 has 6 factors: 12 ms
P7 has 1 factors: 414 ms
P8 has 1 factors: 126 ms
M12_5 has 1 factors: 416 ms
M12_6 has 2 factors: 5838 ms
T1 has 2 factors: 105 ms
T2 has 2 factors: 113 ms
T3 has 4 factors: 3170 ms
H1 has 28 factors: 1998 ms
S7 has 1 factors: 144 ms
S8 has 1 factors: 1623 ms
C1 has 32 factors: 570 ms
```

After:

```
P1 has 36 factors: 15 ms
P2 has 12 factors: 31 ms
P3 has 16 factors: 23 ms
P4 has 2 factors: 188 ms
P5 has 1 factors: 8 ms
P6 has 6 factors: 11 ms
P7 has 1 factors: 415 ms
P8 has 1 factors: 126 ms
M12_5 has 1 factors: 417 ms
M12_6 has 2 factors: 5315 ms
T1 has 2 factors: 106 ms
T2 has 2 factors: 114 ms
T3 has 4 factors: 2250 ms
H1 has 28 factors: 2019 ms
S7 has 1 factors: 145 ms
S8 has 1 factors: 1641 ms
C1 has 32 factors: 573 ms
```