@fredrik-johansson 2024-01-14
It also speeds up the huge_expr example from 3.4 to 2.4 seconds.

There is a surprising failure for Nemo on CI though:

```
fq_nmod_embed.embed: Test Failed at /home/runner/work/flint/flint/Nemo.jl/test/flint/fq_nmod_embed-test.jl:72
  Expression: (f[k8, k16])((f[k4, k8])(x4)) == (f[k4, k16])(x4)
   Evaluated: 642*x16^15 + 98*x16^14 + 363*x16^13 + 73*x16^12 + 894*x16^11 + 859*x16^10 + 276*x16^9 + 811*x16^8 + 77*x16^7 + 519*x16^6 + 836*x16^5 + 693*x16^4 + 594*x16^3 + 736*x16^2 + 531*x16 + 930 == 558*x16^15 + 125*x16^14 + 456*x16^13 + 826*x16^12 + 15*x16^11 + 163*x16^10 + 679*x16^9 + 749*x16^8 + 916*x16^7 + 548*x16^6 + 845*x16^5 + 262*x16^4 + 811*x16^3 + 934*x16^2 + 328*x16 + 472
```